Major change: merging OpenDrift3D and OceanDrift3D into OceanDrift, and likewise OpenOil3D into OpenOil. Updated examples and tests, but documentation remains.
