+++ examples/example_advection_schemes_eddy.py
@@ -26,6 +26,7 @@ for scheme in ['euler', 'runge-kutta', 'runge-kutta4']:
         o.set_config('drift:scheme', scheme)
         o.add_reader(fake_eddy)
         o.seed_elements(lon=2.0, lat=63.0, time=datetime.now())
+        o.set_config('processes:turbulentmixing', False)
         o.run(duration=timedelta(days=9), time_step=time_step)
         runs.append(o)
         i = i + 1
