+++ examples/example_backandforth.py
@@ -24,6 +24,7 @@ o.add_reader(reader_norkyst)
 lon = 4.2; lat = 60.1;
 time = reader_norkyst.start_time
 o.seed_elements(lon, lat, radius=1000, number=100, time=time)
+o.set_config('processes:turbulentmixing', False)
 
 o.run(steps=50*4, time_step=900, outfile=ncfile)
 
@@ -45,6 +46,7 @@ del o
 o = OceanDrift(loglevel=20)  # Set loglevel to 0 for debug information
 o.add_reader(reader_norkyst)
 o.schedule_elements(elements_final, time_final)
+o.set_config('processes:turbulentmixing', False)
 
 #%%
 # Running model backwards from end of forward simulation
