+++ examples/example_cone_tromso.py
@@ -37,6 +37,7 @@ print(o)
 o.set_config('processes:dispersion', True)
 o.set_config('processes:evaporation', False)
 o.set_config('processes:emulsification', True)
+o.set_config('processes:turbulentmixing', False)
 
 #%%
 # Running model for 48 hours
