+++ examples/example_double_gyre_LCS_particles.py
@@ -29,6 +29,7 @@ o.fallback_values['land_binary_mask'] = 0
 #%%
 # Note that Runge-Kutta here makes a difference to Euler scheme
 o.set_config('drift:scheme', 'runge-kutta4')
+o.set_config('processes:turbulentmixing', False)
 
 double_gyre = reader_double_gyre.Reader(epsilon=.25, omega=0.628, A=0.1)
 print(double_gyre)
@@ -50,6 +51,7 @@ lon, lat = double_gyre.xy2lonlat(x, y)
 
 o.seed_elements(lon, lat, radius=.1, number=1000,
                 time=double_gyre.initial_time)
+o.set_config('processes:turbulentmixing', False)
 o.run(duration=duration, time_step=time_step,
       time_step_output=time_step_output)
 o.animation(buffer=0, lcs=lcs)
