+++ examples/example_drifter.py
@@ -41,6 +41,9 @@ o.seed_elements(4.7, 59.9, radius=3000, number=2000,
                 time=reader_current.start_time,
                 wind_drift_factor=wind_drift_factor)
 
+# Prevent mixing elements downwards
+o.set_config('processes:turbulentmixing', False)
+
 #%%
 # Running model
 o.run(time_step=timedelta(minutes=15),
