+++ examples/example_ensemble.py
@@ -13,7 +13,8 @@ from opendrift.readers import reader_netCDF_CF_generic
 # Drift simulation using 10 member ensemble wind data
 # from MEPS model of MET Norway
 
-o = OceanDrift(loglevel=20)
+o = OceanDrift(loglevel=0)
+o.set_config('processes:turbulentmixing', False)
 r = reader_netCDF_CF_generic.Reader('https://thredds.met.no/thredds/dodsC/mepslatest/meps_lagged_6_h_latest_2_5km_latest.nc')
 o.add_reader(r)
