+++ examples/example_grid.py
@@ -15,6 +15,7 @@ reader_norkyst = reader_netCDF_CF_generic.Reader(o.test_data_folder() + '16Nov20
 
 o.add_reader(reader_norkyst)
 #o.fallback_values['land_binary_mask'] = 0
+o.set_config('processes:turbulentmixing', False)
 
 #%%
 # Seeding some particles
