+++ examples/example_oil_thickness.py
@@ -7,7 +7,7 @@ Oil thickness
 from datetime import datetime, timedelta
 import matplotlib.pyplot as plt
 import numpy as np
-from opendrift.models.openoil3D import OpenOil3D
+from opendrift.models.openoil import OpenOil
 
 
 number = 10000
@@ -20,7 +20,7 @@ oiltype = '*GENERIC DIESEL'
 
 #%%
 # First run, where surface oil thickness is updated
-o1 = OpenOil3D(loglevel=30, weathering_model='noaa')
+o1 = OpenOil(loglevel=30, weathering_model='noaa')
 #%%
 # Northwards wind, eastwards current
 o1.fallback_values['land_binary_mask'] = 0
@@ -59,7 +59,7 @@ o1.animation(color='oil_film_thickness', fast=True,
 
 #%%
 # Second run, identical but without updating surface oil thickness
-o2 = OpenOil3D(loglevel=30, weathering_model='noaa')
+o2 = OpenOil(loglevel=30, weathering_model='noaa')
 o2.fallback_values = o1.fallback_values
 
 o2.set_config('wave_entrainment:droplet_size_distribution',
