+++ examples/example_openoil_sample_output.py
@@ -8,9 +8,9 @@ import os
 from datetime import timedelta
 from opendrift.readers import reader_netCDF_CF_generic
 from opendrift.readers import reader_constant
-from opendrift.models.openoil3D import OpenOil3D
+from opendrift.models.openoil import OpenOil
 
-o = OpenOil3D(loglevel=20, weathering_model='noaa')
+o = OpenOil(loglevel=20, weathering_model='noaa')
 
 rc = reader_constant.Reader({'x_wind': 5,
                              'y_wind': 4})
