+++ examples/example_reader_boundary.py
@@ -34,6 +34,8 @@ lons, lats = np.meshgrid(lons, lats)
 o.seed_elements(lons, lats, radius=0, number=2500,
                 time=reader_norkyst.start_time)
 
+o.set_config('processes:turbulentmixing', False)
+
 #%%
 # Running model
 o.run(steps=16*4, time_step=900, time_step_output=1800)
