+++ examples/example_relative.py
@@ -35,6 +35,7 @@ o.seed_elements(lon, lat, radius=50, number=5000, time=time)
 #%%
 # Using windspeed relative to moving ocean (current)
 o.set_config('drift:relative_wind',  False)
+o.set_config('processes:turbulentmixing', False)
 
 o.run(steps=48*2, time_step=1800, time_step_output=3600*2)
 
@@ -44,6 +45,7 @@ o2 = OceanDrift(loglevel=20)  # Set loglevel to 0 for debug information
 o2.add_reader([reader_norkyst, reader_arome])
 o2.seed_elements(lon, lat, radius=50, number=5000, time=time)
 o2.set_config('drift:relative_wind',  True)
+o2.set_config('processes:turbulentmixing', False)
 o2.run(steps=48*2, time_step=1800, time_step_output=3600*2)
