+++ examples/example_satellite.py
@@ -26,6 +26,7 @@ o.seed_from_gml(o.test_data_folder() + 'radarsat_oil_satellite_observation/RS2_2
 
 #%%
 # Adjusting some configuration
+o.set_config('processes:turbulentmixing', False)
 o.set_config('processes:dispersion', True)
 o.set_config('processes:evaporation', False)
 o.set_config('processes:emulsification', True)
