--- opendrift/models/oceandrift3D.py
@@ -1,120 +1,0 @@
-# This file is part of OpenDrift.
-#
-# OpenDrift is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, version 2
-#
-# OpenDrift is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with OpenDrift.  If not, see <https://www.gnu.org/licenses/>.
-#
-# Copyright 2015, Knut-Frode Dagestad, MET Norway
-
-import numpy as np
-from opendrift.models.opendrift3D import OpenDrift3DSimulation, Lagrangian3DArray
-from opendrift.models.oceandrift import OceanDrift
-
-# We add the property 'wind_drift_factor' to the element class
-Lagrangian3DArray.variables = Lagrangian3DArray.add_variables([
-                            ('wind_drift_factor', {'dtype': np.float32,
-                                                   'unit': '%',
-                                                   'default': 0.0})])
-
-
-class OceanDrift3D(OpenDrift3DSimulation, OceanDrift):
-    """Trajectory model based on the OpenDrift framework.
-
-    Simply propagation with horizontal and vertical ocean currents
-    and possibly additional wind drag.
-    Suitable for passive tracers, e.g. for tracking water particles.
-    Developed at MET Norway.
-
-    """
-
-    ElementType = Lagrangian3DArray
-    required_variables = [
-        'x_sea_water_velocity',
-        'y_sea_water_velocity',
-        'x_wind', 'y_wind',
-        'upward_sea_water_velocity',
-        'ocean_vertical_diffusivity',
-        'sea_surface_wave_significant_height',
-        'sea_surface_wave_stokes_drift_x_velocity',
-        'sea_surface_wave_stokes_drift_y_velocity',
-        'sea_surface_wave_period_at_variance_spectral_density_maximum',
-        'sea_surface_wave_mean_period_from_variance_spectral_density_second_frequency_moment',
-        'surface_downward_x_stress',
-        'surface_downward_y_stress',
-        'turbulent_kinetic_energy',
-        'turbulent_generic_length_scale',
-        'sea_floor_depth_below_sea_level',
-        'land_binary_mask'
-        ]
-
-    required_profiles = ['ocean_vertical_diffusivity']
-    # The depth range (in m) which profiles shall cover
-    required_profiles_z_range = [-120, 0]
-
-    fallback_values = {
-        'x_sea_water_velocity': 0,
-        'y_sea_water_velocity': 0,
-        'sea_surface_wave_significant_height': 0,
-        'sea_surface_wave_stokes_drift_x_velocity': 0,
-        'sea_surface_wave_stokes_drift_y_velocity': 0,
-        'sea_surface_wave_period_at_variance_spectral_density_maximum': 0,
-        'sea_surface_wave_mean_period_from_variance_spectral_density_second_frequency_moment': 0,
-        'x_wind': 0,
-        'y_wind': 0,
-        'upward_sea_water_velocity': 0,
-        'ocean_vertical_diffusivity': 0.02,
-        'surface_downward_x_stress': 0,
-        'surface_downward_y_stress': 0,
-        'turbulent_kinetic_energy': 0,
-        'turbulent_generic_length_scale': 0,
-        'sea_floor_depth_below_sea_level': 10000
-        }
-
-    def __init__(self, *args, **kwargs):
-
-        # Calling general constructor of parent class
-        super(OceanDrift3D, self).__init__(*args, **kwargs)
-
-        # Turbulent mixing switched off by default
-        self.set_config('processes:turbulentmixing', False)
-
-    def update_terminal_velocity(self, *args, **kwargs):
-        '''
-        Terminal velocity due to buoyancy or sedimentation rate,
-        to be used in turbulent mixing module.
-        Using zero for passive particles, i.e. following water particles
-        '''
-        
-        # TODO: setting to scalar has lead to trouble - should be array of num_elements
-        #self.elements.terminal_velocity = 0.
-
-    def update(self):
-        """Update positions and properties of elements."""
-
-        # Simply move particles with ambient current
-        self.advect_ocean_current()
-
-        # Advect particles due to wind drag
-        # (according to specified wind_drift_factor)
-        self.advect_wind()
-
-        # Stokes drift
-        self.stokes_drift()
-
-        # Turbulent Mixing
-        if self.get_config('processes:turbulentmixing') is True:
-            self.update_terminal_velocity()
-            self.vertical_mixing()
-        else:  # Buoyancy
-            self.vertical_buoyancy()
-
-        # Vertical advection
-        self.vertical_advection()
