+++ opendrift/models/openoil.py
@@ -62,20 +62,28 @@ The droplet size limits could also have been calculated dynamically from oil vis
 
 from io import open
 import os
+import json
 import numpy as np
 from datetime import datetime
 import pyproj
 import matplotlib.pyplot as plt
 
-from opendrift.models.basemodel import OpenDriftSimulation
+#from opendrift.models.basemodel import OpenDriftSimulation
+from opendrift.models.oceandrift import OceanDrift
 from opendrift.elements import LagrangianArray
 import opendrift.models.noaa_oil_weathering as noaa
+from opendrift.models.physics_methods import oil_wave_entrainment_rate_li2017, oil_wave_entrainment_rate_tkalich2002
+
 
 try:
     from itertools import izip as zip
 except ImportError:
     pass
 
+try:
+    basestring
+except NameError:
+    basestring = str
 
 # Defining the oil element properties
 class Oil(LagrangianArray):
@@ -124,10 +132,17 @@ class Oil(LagrangianArray):
                             'default': 0}),
         ('oil_film_thickness', {'dtype': np.float32,
                                 'units': 'm',
-                                'default': 0.001})])
+                                'default': 0.001}),
+        ('entrainment_length_scale', {'dtype': np.float32,
+                                      'units': 'm',
+                                      'default': 0.03}),
+        ('diameter', {'dtype': np.float32,  # Particle diameter
+                      'units': 'm',
+                      'default': 0.})
+        ])
 
 
-class OpenOil(OpenDriftSimulation):
+class OpenOil(OceanDrift):
     """Open source oil trajectory model based on the OpenDrift framework.
 
         Developed at MET Norway based on oil weathering parameterisations
@@ -139,43 +154,78 @@ class OpenOil(OpenDriftSimulation):
     ElementType = Oil
 
     required_variables = ['x_sea_water_velocity', 'y_sea_water_velocity',
+                          'upward_sea_water_velocity',
                           'sea_surface_wave_significant_height',
                           'sea_surface_wave_stokes_drift_x_velocity',
                           'sea_surface_wave_stokes_drift_y_velocity',
+                          'sea_surface_wave_period_at_variance_spectral_density_maximum',
+                          'sea_surface_wave_mean_period_from_variance_spectral_density_second_frequency_moment',
                           'sea_ice_area_fraction',
                           'sea_ice_x_velocity', 'sea_ice_y_velocity',
                           'sea_water_temperature',
+                          'sea_water_salinity',
                           'sea_floor_depth_below_sea_level',
-                          'x_wind', 'y_wind', 'land_binary_mask']
+                          'x_wind', 'y_wind',
+                          'ocean_vertical_diffusivity',
+                          'land_binary_mask']
 
-    fallback_values = {'x_sea_water_velocity': 0,
-                       'y_sea_water_velocity': 0,
+    # Desired variables do not require initialisation of Lazy readers
+    desired_variables = [
+        'sea_surface_wave_significant_height',
+        'sea_surface_wave_stokes_drift_x_velocity',
+        'sea_surface_wave_stokes_drift_y_velocity',
+        'sea_surface_wave_period_at_variance_spectral_density_maximum',
+        'sea_surface_wave_mean_period_from_variance_spectral_density_second_frequency_moment',
+        'sea_ice_area_fraction',
+        'sea_ice_x_velocity', 'sea_ice_y_velocity',
+        'ocean_vertical_diffusivity',
+        'upward_sea_water_velocity'
+        ]
+
+
+    fallback_values = {#'x_sea_water_velocity': 0,
+                       #'y_sea_water_velocity': 0,
+                       'upward_sea_water_velocity': 0,
+                       #'x_wind': 0, 'y_wind': 0,
                        'sea_surface_wave_significant_height': 0,
                        'sea_surface_wave_stokes_drift_x_velocity': 0,
                        'sea_surface_wave_stokes_drift_y_velocity': 0,
-                       'sea_floor_depth_below_sea_level': 0,
+                       'sea_surface_wave_period_at_variance_spectral_density_maximum': 0,
+        'sea_surface_wave_mean_period_from_variance_spectral_density_second_frequency_moment': 0,
+                       'sea_floor_depth_below_sea_level': 10000,
                        'sea_ice_area_fraction': 0,
                        'sea_ice_x_velocity': 0,
                        'sea_ice_y_velocity': 0,
-                       'sea_water_temperature': 12,
-                       'x_wind': 0, 'y_wind': 0}
+                       'sea_water_temperature': 10,
+                       'sea_water_salinity': 34.,
+                       'ocean_vertical_diffusivity': 0.02  # m2s-1
+                        }
+
+    required_profiles = ['sea_water_temperature',
+                         'sea_water_salinity',
+                         'ocean_vertical_diffusivity']
+    # The depth range (in m) which profiles shall cover
+    required_profiles_z_range = [-120, 0]
+
+
+    max_speed = 1.3  # m/s
 
     # Default colors for plotting
     status_colors = {'initial': 'green', 'active': 'blue',
                      'missing_data': 'gray', 'stranded': 'red',
                      'evaporated': 'yellow', 'dispersed': 'magenta'}
 
-    configspec = '''
-        [processes]
-            dispersion = boolean(default=True)
-            evaporation = boolean(default=True)
-            emulsification = boolean(default=True)
-            biodegradation = boolean(default=False)
-            update_oilfilm_thickness = boolean(default=False)
-        [drift]
-            current_uncertainty = float(min=0, max=5, default=0.05)
-            wind_uncertainty = float(min=0, max=5, default=.5)
-    '''
+#    configspec = '''
+#        [processes]
+#            dispersion = boolean(default=True)
+#            evaporation = boolean(default=True)
+#            emulsification = boolean(default=True)
+#            biodegradation = boolean(default=False)
+#            update_oilfilm_thickness = boolean(default=False)
+#        [drift]
+#            current_uncertainty = float(min=0, max=5, default=0.05)
+#            wind_uncertainty = float(min=0, max=5, default=.5)
+#    '''
 
     duplicate_oils = ['ALVHEIM BLEND, STATOIL', 'DRAUGEN, STATOIL',
                       'EKOFISK BLEND 2000', 'EKOFISK BLEND, STATOIL',
@@ -197,6 +247,40 @@ class OpenOil(OpenDriftSimulation):
     max_water_fraction  = {
         'MARINE GAS OIL 500 ppm S 2017': 0.1,
         'FENJA (PIL) 2015': .75}
+
+    # Read oil types from file (presently only for illustrative effect)
+    oil_types = str([str(l.strip()) for l in open(
+                    os.path.dirname(os.path.realpath(__file__)) +
+                    '/oil_types.txt').readlines()])[1:-1]
+    default_oil = oil_types.split(',')[0].strip()
+
+    # Configuration
+    configspecOO = '''
+        [input]
+            [[spill]]
+                oil_type = option(%s, default=%s)
+                droplet_diameter_min_subsea = float(min=1e-8, max=1, default=0.0005)
+                droplet_diameter_max_subsea = float(min=1e-8, max=1, default=0.005)
+        [processes]
+            dispersion = boolean(default=True)
+            evaporation = boolean(default=True)
+            emulsification = boolean(default=True)
+            biodegradation = boolean(default=False)
+            update_oilfilm_thickness = boolean(default=False)
+            turbulentmixing = boolean(default=True)
+        [drift]
+            wind_drift_depth = float(min=0, max=10, default=0.1)
+            verticaladvection = boolean(default=False)
+            current_uncertainty = float(min=0, max=5, default=0.05)
+            wind_uncertainty = float(min=0, max=5, default=.5)
+        [wave_entrainment]
+            droplet_size_distribution = option('Exponential', 'Johansen et al. (2015)', 'Li et al. (2017)', default='Johansen et al. (2015)')
+            entrainment_rate = option('Tkalich & Chan (2002)', 'Li et al. (2017)', default='Li et al. (2017)')
+        [turbulentmixing]
+            droplet_diameter_min_wavebreaking = float(default=1e-5, min=1e-8, max=1)
+            droplet_diameter_max_wavebreaking = float(default=2e-3, min=1e-8, max=1)
+            droplet_size_exponent = float(default=0, min=-10, max=10)
+    ''' % (oil_types, default_oil) 
 
 
     def __init__(self, weathering_model='default', *args, **kwargs):
@@ -250,7 +334,7 @@ class OpenOil(OpenDriftSimulation):
         super(OpenOil, self).__init__(*args, **kwargs)
 
         # Overriding with specific configspec
-        self._add_configstring(self.configspec)
+        self._add_configstring(self.configspecOO)
 
     def update_surface_oilfilm_thickness(self):
         '''The mass of oil is summed within a grid of 20x20
@@ -480,6 +564,19 @@ class OpenOil(OpenDriftSimulation):
         # Biodegradation
         self.biodegradation()
 
+    def particle_radius(self):
+        """Calculate radius of entained particles.
+
+        Per now a fixed radius, should later use a distribution.
+        """
+
+        # Delvigne and Sweeney (1988)
+        # rmax = 1818*np.power(self.wave_energy_dissipation(), -0.5)* \
+        #             np.power(self.elements.viscosity, 0.34) / 1000000
+
+        # r = np.random.uniform(0, rmax, self.num_elements_active())
+        return self.elements.diameter/2.0  # Hardcoded diameter
+
     def prepare_run(self):
 
         if self.oil_weathering_model == 'noaa':
@@ -679,6 +776,264 @@ class OpenOil(OpenDriftSimulation):
         self.elements.water_fraction[self.elements.interfacial_area >=
             ((6.0 / drop_max)*(Y_max/(1.0 - Y_max)))] = Y_max
 
+
+    def update_terminal_velocity(self, Tprofiles=None,
+                                 Sprofiles=None, z_index=None):
+        """Calculate terminal velocity for oil droplets
+
+        according to
+        Tkalich et al. (2002): Vertical mixing of oil droplets
+                               by breaking waves
+        Marine Pollution Bulletin 44, 1219-1229
+
+        If profiles of temperature and salt are passed into this function,
+        they will be interpolated from the profiles.
+        if not, T,S will be fetched from reader.
+        """
+        g = 9.81  # ms-2
+
+        r = self.particle_radius()*2.0
+
+        # prepare interpolation of temp, salt
+
+        if not (Tprofiles is None and Sprofiles is None):
+            if z_index is None:
+                z_i = range(Tprofiles.shape[0])  # evtl. move out of loop
+                # evtl. move out of loop
+                z_index = interp1d(-self.environment_profiles['z'],
+                                   z_i, bounds_error=False)
+            zi = z_index(-self.elements.z)
+            upper = np.maximum(np.floor(zi).astype(np.int), 0)
+            lower = np.minimum(upper+1, Tprofiles.shape[0]-1)
+            weight_upper = 1 - (zi - upper)
+
+        # do interpolation of temp, salt if profiles were passed into
+        # this function, if not, use reader by calling self.environment
+        if Tprofiles is None:
+            T0 = self.environment.sea_water_temperature
+        else:
+            T0 = Tprofiles[upper, range(Tprofiles.shape[1])] * \
+                weight_upper + \
+                Tprofiles[lower, range(Tprofiles.shape[1])] * \
+                (1-weight_upper)
+        if Sprofiles is None:
+            S0 = self.environment.sea_water_salinity
+        else:
+            S0 = Sprofiles[upper, range(Sprofiles.shape[1])] * \
+                weight_upper + \
+                Sprofiles[lower, range(Sprofiles.shape[1])] * \
+                (1-weight_upper)
+
+        rho_oil = self.elements.density
+        rho_water = self.sea_water_density(T=T0, S=S0)
+
+        # dynamic water viscosity
+        my_w = 0.001*(1.7915 - 0.0538*T0 + 0.007*(T0**(2.0)) - 0.0023*S0)
+        # ~0.0014 kg m-1 s-1
+        # kinemativ water viscosity
+        ny_w = my_w / rho_water
+        rhopr = rho_oil/rho_water
+
+        # terminal velocity for low Reynolds numbers
+        kw = 2*g*(1-rhopr)/(9*ny_w)
+        W = kw * r**2
+        
+        # check if we are in a high Reynolds number regime
+        Re = 2*r*W/ny_w
+        highRe = np.where(Re > 50)
+
+        # Terminal velocity in high Reynolds numbers
+        kw = (16*g*(1-rhopr)/3)**0.5
+        W2 = kw*r**0.5
+
+        W[highRe] = W2[highRe]
+        self.elements.terminal_velocity = W
+
+
+    def oil_wave_entrainment_rate(self):
+        er = self.get_config('wave_entrainment:entrainment_rate')
+        if er == 'Tkalich & Chan (2002)':
+            #entrainment_rate = self.oil_wave_entrainment_rate_tkalich2002()
+            entrainment_rate = oil_wave_entrainment_rate_tkalich2002(
+                    wind_speed=self.wind_speed(),
+                    significant_wave_height=self.significant_wave_height(),
+                    entrainment_length_scale=
+                        self.elements.entrainment_length_scale)
+        elif er == 'Li et al. (2017)':
+            entrainment_rate = oil_wave_entrainment_rate_li2017(
+                    dynamic_viscosity=self.elements.viscosity*
+                        self.elements.density,
+                    oil_density=self.elements.density,
+                    interfacial_tension=self.oil_water_interfacial_tension,
+                    significant_wave_height=self.significant_wave_height(),
+                    wave_breaking_fraction=self.sea_surface_wave_breaking_fraction(),
+                    sea_water_density=self.sea_water_density())
+        return entrainment_rate
+
+
+    def prepare_vertical_mixing(self):
+        '''Calculate entrainment probability before main loop'''
+        self.oil_entrainment_probability = \
+            1 - np.exp(-self.oil_wave_entrainment_rate()*\
+                       self.get_config('turbulentmixing:timestep'))
+        # Calculate a random droplet diameter for each particle,
+        # to be used if this particle gets entrained
+        self.droplet_diameter_if_entrained = \
+            self.get_wave_breaking_droplet_diameter()
+        # Uncomment lines below to plot droplet size distribution at each step
+        #import matplotlib.pyplot as plt
+        #plt.hist(self.droplet_diameter_if_entrained, 200)
+        #plt.gca().set_xscale("log")
+        #plt.gca().set_yscale("log")
+        #plt.show()
+
+
+    def surface_wave_mixing(self, time_step_seconds):
+        """Mix surface oil into water column."""
+        # Entrain oil into uppermost layer (whitecapping from waves)
+        # TODO: optimise this by only calculate for surface elements
+        surface = self.elements.z >= 0
+        random_number = np.random.uniform(0, 1, len(self.elements.z))
+        entrained = np.logical_and(surface,
+                        random_number<self.oil_entrainment_probability)
+
+        # Intrusion depth for wave entrainment from Delvigne and Sweeney (1988), Li et al. (2017):
+        if entrained.sum() > 0:
+            self.logger.debug('Entraining %i of %i surface elements' %
+                          (entrained.sum(), surface.sum()))
+            zb = 1.5 * self.significant_wave_height() # between 0 and zb
+            intrusion_depth = np.random.uniform(0, np.mean(zb), entrained.sum())
+            self.elements.z[entrained] = - intrusion_depth
+            if self.keep_droplet_diameter is False:
+                # Give entrained elements a random diameter
+                self.elements.diameter[entrained] = \
+                    self.droplet_diameter_if_entrained[entrained]
+
+    def surface_stick(self):
+        """set surfaced particles to exactly zero depth to let them form a slick """
+
+        surface = np.where(self.elements.z >= 0)
+        if len(surface[0]) > 0:
+            self.elements.z[surface] = 0.
+
+    def get_wave_breaking_droplet_diameter(self):
+        dm = self.get_config('wave_entrainment:droplet_size_distribution')
+        if dm == 'Johansen et al. (2015)':
+            d = self.get_wave_breaking_droplet_diameter_johansen2015()
+        elif dm == 'Li et al. (2017)':
+            d = self.get_wave_breaking_droplet_diameter_liz2017()
+        elif dm == 'Exponential':
+            d = self.get_wave_breaking_droplet_diameter_exponential()
+        return d
+
+    def get_wave_breaking_droplet_diameter_exponential(self):
+        if not hasattr(self, 'droplet_spectrum_pdf'):
+            # Generate droplet spectrum, if not already done
+            self.logger.debug('Generating wave breaking droplet size spectrum')
+            s = self.get_config('turbulentmixing:droplet_size_exponent')
+            dmax = self.get_config('turbulentmixing:droplet_diameter_max_wavebreaking')
+            dmin = self.get_config('turbulentmixing:droplet_diameter_min_wavebreaking')
+            # Note: a long array of diameters is required for 
+            # sufficient resolution at both ends of logarithmic scale.
+            # Could perhaps use logspace instead of linspace(?)
+            self.droplet_spectrum_diameter = np.linspace(dmin, dmax, 1000000)
+            spectrum = self.droplet_spectrum_diameter**s
+            self.droplet_spectrum_pdf = spectrum/np.sum(spectrum)
+
+        return np.random.choice(self.droplet_spectrum_diameter,
+                                size=self.num_elements_active(),
+                                p=self.droplet_spectrum_pdf)
+
+    def get_wave_breaking_droplet_diameter_liz2017(self):
+        # Li,Zhengkai, M. Spaulding, D. French-McCay, D. Crowley, J.R. Payne: "Development of a unified oil droplet size distribution model 
+        # with application to surface breaking waves and subsea blowout releases considering dispersant effects" Mar. Pol. Bul.
+        # DOI: 10.1016/j.marpolbul.2016.09.008
+        # Should be prefered when the oil film thickness is unknown.
+        if not hasattr(self, 'droplet_spectrum_pdf'):
+            # Generate droplet spectrum as in Li (Zhengkai) et al. (2017)
+            self.logger.debug('Generating wave breaking droplet size spectrum')
+            dmax = self.get_config('turbulentmixing:droplet_diameter_max_wavebreaking')
+            dmin = self.get_config('turbulentmixing:droplet_diameter_min_wavebreaking')
+            self.droplet_spectrum_diameter = np.linspace(dmin, dmax, 1000000)
+            g = 9.81
+            interfacial_tension = self.oil_water_interfacial_tension
+            delta_rho = self.sea_water_density() - self.elements.density
+            d_o = 4 * (interfacial_tension / (delta_rho*g))**0.5
+            we = ( self.sea_water_density() * g * self.significant_wave_height() * d_o ) / interfacial_tension
+            oh = self.elements.viscosity * self.elements.density * (self.elements.density * interfacial_tension * d_o )**-0.5 # From kin. to dyn. viscosity by * density
+            r = 1.791
+            p = 0.460
+            q = -0.518
+            dV_50 = d_o * r * (1+10*oh)**p * we**q # median droplet diameter in volume distribution
+            sd = 0.4 # log standard deviation in log10 units
+            Sd = np.log(10) *sd # log standard deviation in natural log units
+            # TODO: calculation below with scalars, but we have arrays, with varying oil properties
+            # treat all particle in one go:
+            dV_50 = np.mean(dV_50) # mean log diameter
+            dN_50 = np.exp( np.log(dV_50) - 3*Sd**2 ) # convert number distribution to volume distribution
+            self.logger.debug('Droplet distribution median diameter dV_50: %f, dN_50: %f ' %( dV_50, np.mean(dN_50)))
+            spectrum = (np.exp(-(np.log(self.droplet_spectrum_diameter) - np.log(dV_50))**2 / (2 * Sd**2))) / (self.droplet_spectrum_diameter * Sd * np.sqrt(2 * np.pi))
+            self.droplet_spectrum_pdf = spectrum/np.sum(spectrum)
+        if ~np.isfinite(np.sum(self.droplet_spectrum_pdf)) or \
+                np.abs(np.sum(self.droplet_spectrum_pdf) - 1) > 1e-6:
+            self.logger.warning('Could not update droplet diameters.')
+            return self.elements.diameter
+        else:
+            return np.random.choice(self.droplet_spectrum_diameter,
+                                    size=self.num_elements_active(),
+                                    p=self.droplet_spectrum_pdf)
+
+    def get_wave_breaking_droplet_diameter_johansen2015(self):
+        # Johansen O, Reed M, Bodsberg NR, Natural dispersion revisited
+        # DOI: 10.1016/j.marpolbul.2015.02.026
+        # requires oil film thickness
+        if not hasattr(self, 'droplet_spectrum_pdf') or self.get_config('processes:update_oilfilm_thickness') is True:
+            # Generate droplet spectrum as in Johansen et al. (2015)
+            self.logger.debug('Generating wave breaking droplet size spectrum')
+            dmax = self.get_config('turbulentmixing:droplet_diameter_max_wavebreaking')
+            dmin = self.get_config('turbulentmixing:droplet_diameter_min_wavebreaking')
+            self.droplet_spectrum_diameter = np.linspace(dmin, dmax, 1000000)
+            g = 9.81
+            interfacial_tension = self.oil_water_interfacial_tension
+            #
+            #A = self.significant_wave_height()/2. # wave amplitude
+            #re = (self.elements.density*self.elements.oil_film_thickness*(2*g*A)**0.5) / (self.elements.viscosity*self.elements.density) # Reyolds number
+            #we = (self.elements.density*self.elements.oil_film_thickness*2*g*A) / interfacial_tension # Weber number
+            #
+            H = self.significant_wave_height() # fall height = 2 * wave amplitude
+            # Reyolds number (Eq. 7a from Johansen et al. 2015)
+            re = (self.elements.density*self.elements.oil_film_thickness*(g*H)**0.5) / (self.elements.viscosity*self.elements.density)
+            # Weber number (Eq. 7b from Johansen et al.2015)
+            we = (self.elements.density*self.elements.oil_film_thickness*g*H) / interfacial_tension # Weber number
+            A = 2.251 # parameters from Johansen et al. 2015
+            Bp = 0.027
+            B = A*Bp
+            dN_50 = (A*self.elements.oil_film_thickness*we**-0.6) + (B*self.elements.oil_film_thickness* re**-0.6) # median droplet diameter in number distribution
+            sd = 0.4 # log standard deviation in log10 units
+            Sd = np.log(10) *sd # log standard deviation in natural log units
+            dV_50 = np.exp( np.log(dN_50) + 3*Sd**2 ) # convert number distribution to volume distribution
+            # TODO: calculation below with scalars, but we have
+            # arrays, with varying oil properties
+            # treat all particle in one go:
+            dV_50 = np.mean(dV_50) # mean log diameter
+            self.logger.debug('Droplet distribution median diameter dV_50: %f, dN_50: %f ' %( dV_50, np.mean(dN_50)))
+            spectrum = (np.exp(-(np.log(self.droplet_spectrum_diameter) - np.log(dV_50))**2 / (2 * Sd**2))) / (self.droplet_spectrum_diameter * Sd * np.sqrt(2 * np.pi))
+            self.droplet_spectrum_pdf = spectrum/np.sum(spectrum)
+        if ~np.isfinite(np.sum(self.droplet_spectrum_pdf)) or \
+                np.abs(np.sum(self.droplet_spectrum_pdf) - 1) > 1e-6:
+            self.logger.warning('Could not update droplet diameters.')
+            return self.elements.diameter
+        else:
+            return np.random.choice(self.droplet_spectrum_diameter,
+                                    size=self.num_elements_active(),
+                                    p=self.droplet_spectrum_pdf)
+
+    def resurface_elements(self, minimum_depth=None):
+        """Oil elements reaching surface (or above) form slick, not droplet"""
+        surface = np.where(self.elements.z >= 0)[0]
+        self.elements.z[surface] = 0
+
+
     def advect_oil(self):
 
         # Calculating various drift factors according to ice concentration
@@ -722,14 +1077,37 @@ class OpenOil(OpenDriftSimulation):
         #        self.environment.sea_ice_area_fraction > 0.6,
         #        reason='oil-in-ice')
 
+    #def update(self):
+    #    """Update positions and properties of oil particles."""
+
+    #    # Oil weathering
+    #    self.oil_weathering()
+
+    #    # Horizontal advection
+    #    self.advect_oil()
+
     def update(self):
         """Update positions and properties of oil particles."""
 
-        # Oil weathering
+        if self.get_config('processes:update_oilfilm_thickness') is True:
+            self.update_surface_oilfilm_thickness()
+
+        # Oil weathering (inherited from OpenOil)
         self.oil_weathering()
 
-        # Horizontal advection
+        # Turbulent Mixing
+        if self.get_config('processes:turbulentmixing') is True:
+            self.update_terminal_velocity()
+            self.vertical_mixing()
+            del self.droplet_spectrum_pdf
+
+        # Vertical advection
+        if self.get_config('processes:verticaladvection') is True:
+            self.vertical_advection()
+
+        # Horizontal advection (inherited from OpenOil)
         self.advect_oil()
+
 
     def get_oil_budget(self):
         """Get oil budget for the current simulation
@@ -1013,6 +1391,76 @@ class OpenOil(OpenDriftSimulation):
         self.oil_data['oiltype'] = oiltype  # Store name of oil type
 
     def seed_elements(self, *args, **kwargs):
+
+        # Old OpenOil3D seeding code
+        if len(args) == 2:
+            kwargs['lon'] = args[0]
+            kwargs['lat'] = args[1]
+            args = {}
+
+        seed_json = {'start':{}, 'end':{}}
+        for kw in kwargs:
+            data = kwargs[kw]
+            if not isinstance(data, basestring):
+                data = np.atleast_1d(data)
+            if isinstance(data[0], datetime):
+                data[0] = str(data[0])
+                if len(data) == 2:
+                    data[1] = str(data[1])
+            if not isinstance(kwargs[kw], basestring):
+                if kw in ['lon', 'lat', 'z', 'radius', 'time']:
+                    pointer = seed_json['start']
+                    pointer2 = seed_json['end']
+                else:
+                    pointer = seed_json
+                if len(data) == 1:
+                    self.add_metadata('seed_' + kw, data[0])
+                    pointer[kw] = data[0]
+                elif len(kwargs[kw]) == 2:
+                    self.add_metadata('seed_' + kw + '_start', str(kwargs[kw][0]))
+                    self.add_metadata('seed_' + kw + '_end', str(kwargs[kw][1]))
+                    pointer[kw] = data[0]
+                    pointer2[kw] = data[1]
+                else:
+                    pass
+                    #self.logger.info('Not adding array %s to metadata' % kw)
+            else:
+                self.add_metadata('seed_' + kw, str(data))
+                seed_json[kw] = data
+
+        if 'number' not in kwargs:
+            number = 1
+        else:
+            number = kwargs['number']
+        if 'diameter' in kwargs:
+            self.logger.info('Droplet diameter is provided, and will '
+                         'be kept constant during simulation')
+            self.keep_droplet_diameter = True
+        else:
+            self.keep_droplet_diameter = False
+        if 'z' not in kwargs:
+            kwargs['z'] = 0
+        if isinstance(kwargs['z'], basestring) and \
+                kwargs['z'][0:8] == 'seafloor':
+            z = -np.ones(number)
+        else:
+            z = np.atleast_1d(kwargs['z'])
+        if len(z) == 1:
+            z = z*np.ones(number)  # Convert scalar z to array
+        subsea = z < 0
+        if np.sum(subsea) > 0 and 'diameter' not in kwargs:
+            # Droplet min and max for particles seeded below sea surface
+            sub_dmin = self.get_config('input:spill:droplet_diameter_min_subsea')
+            sub_dmax = self.get_config('input:spill:droplet_diameter_max_subsea')
+            self.logger.info('Using particle diameters between %s and %s m for '
+                         'elements seeded below sea surface.' %
+                         (sub_dmin, sub_dmax))
+            kwargs['diameter'] = np.random.uniform(sub_dmin, sub_dmax, number)
+
+
+        ##########################
+        # Old OpenOil seeding
+        
         if 'oiltype' in kwargs:
             self.set_config('seed:oil_type', kwargs['oiltype'])
             del kwargs['oiltype']
@@ -1048,6 +1496,52 @@ class OpenOil(OpenDriftSimulation):
             del kwargs['m3_per_hour']
 
         super(OpenOil, self).seed_elements(*args, **kwargs)
+
+        # Add oil metadata
+        try:
+            self.add_metadata('seed_oil_density', self.oiltype.get_density())
+            seed_json['oil_density'] = self.oiltype.get_density()
+        except:
+            try:
+                self.add_metadata('seed_oil_density',
+                                  self.oiltype.density_at_temp(283))
+                seed_json['oil_density'] = self.oiltype.density_at_temp(283)
+            except:
+                pass
+        try:
+            self.add_metadata('seed_oil_viscosity',
+                              self.oiltype.get_viscosity(283))
+            seed_json['oil_viscosity'] = self.oiltype.get_viscosity(283)
+        except:
+            try:
+                self.add_metadata('seed_oil_viscosity',
+                                  self.oiltype.kvis_at_temp(283))
+                seed_json['oil_viscosity'] = self.oiltype.kvis_at_temp(283)
+            except:
+                pass
+
+        if not hasattr(self, 'seed_json'):
+            self.seed_json = []
+        self.seed_json.append(seed_json)
+
+        class MyEncoder(json.JSONEncoder):  # Serialisation of json
+            def default(self, obj):
+                if isinstance(obj, np.bool_) :
+                    return str(obj)
+                elif isinstance(obj, bool) :
+                    return str(obj)
+                elif isinstance(obj, np.integer):
+                    return int(obj)
+                elif isinstance(obj, np.floating):
+                    return float(obj)
+                elif isinstance(obj, np.ndarray):
+                    return obj.tolist()
+                else:
+                    return super(MyEncoder, self).default(obj)
+
+        self.add_metadata('seed_json', json.dumps(self.seed_json,
+                                            cls=MyEncoder))
+
 
     def seed_from_gml(self, gmlfile, num_elements=1000, *args, **kwargs):
         """Read oil slick contours from GML file, and seed particles within."""
