+++ opendrift/models/pelagicegg.py
@@ -16,9 +16,8 @@
 
 import numpy as np
 
-from opendrift.models.opendrift3D import \
-    OpenDrift3DSimulation, Lagrangian3DArray
-from opendrift.elements import LagrangianArray
+from opendrift.models.oceandrift import OceanDrift, Lagrangian3DArray
+#from opendrift.elements import LagrangianArray
 
 
 # Defining the oil element properties
@@ -26,7 +25,7 @@ class PelagicEgg(Lagrangian3DArray):
     """Extending Lagrangian3DArray with specific properties for pelagic eggs
     """
 
-    variables = LagrangianArray.add_variables([
+    variables = Lagrangian3DArray.add_variables([
         ('diameter', {'dtype': np.float32,
                       'units': 'm',
                       'default': 0.0014}),  # for NEA Cod
@@ -41,7 +40,7 @@ class PelagicEgg(Lagrangian3DArray):
                      'default': 0.})])
 
 
-class PelagicEggDrift(OpenDrift3DSimulation):
+class PelagicEggDrift(OceanDrift):
     """Buoyant particle trajectory model based on the OpenDrift framework.
 
         Developed at MET Norway
